fast-agent/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.18.2
python-dotenv>=0.19.1
streamlit>=1.10.0
pydantic>=1.9.0
requests>=2.26.0
jinja2==3.1.2
langchain>=0.0.294
langchain-core>=0.1.4
langchain-google-vertexai>=0.0.1
google-cloud-aiplatform>=1.36.0
markdown>=3.5
python-multipart==0.0.6
starlette>=0.27.0 